+++ nfc_reader/extns/impl/nxpnfc/2.0/NxpNfc.h
@@ -1,6 +1,6 @@
 /******************************************************************************
  *
- *  Copyright 2018-2020,2024 NXP
+ *  Copyright 2018-2020,2024,2026 NXP
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@
 #include <hidl/Status.h>
 #include <vendor/nxp/nxpnfc/2.0/INxpNfc.h>
 
-enum Constants : uint16_t {
+enum Constants : uint8_t {
   HAL_NFC_ESE_HARD_RESET = 5,
 };
